99% done the header
