+++ .prettierrc
@@ -1,6 +1,9 @@
 {
-	"useTabs": false,
-	"singleQuote": true,
-	"trailingComma": "es5",
-	"printWidth": 100
+  "useTabs": false,
+  "singleQuote": true,
+  "trailingComma": "es5",
+  "printWidth": 100,
+  "semi": false,
+  "bracketSpacing": true,
+  "tabWidth": 2
 }
