+++ package.json
@@ -22,6 +22,7 @@
     "prettier-plugin-svelte": "^2.4.0",
     "svelte": "^3.44.0",
     "svelte-check": "^2.2.6",
+    "svelte-feather-icons": "^3.5.1",
     "svelte-preprocess": "^4.9.4",
     "tailwindcss": "^3.0.23",
     "tslib": "^2.3.1",
